[feat] move model to common package
