+++ chalicelib/middleware/error_handler.py
@@ -9,7 +9,7 @@ from chalice.app import (
     Response,
 )
 
-from chalicelib.model.builder import ApiBuilder
+from chalicelib.common.model.builder import ApiBuilder
 
 
 logger = logging.getLogger("pyoniverse-api")
