+++ tests/brand/test_api.py
@@ -1,5 +1,5 @@
 from chalicelib.domain.brand.model.brand_response import BrandDetailResponseSchema
-from chalicelib.model.api import ApiSchema
+from chalicelib.common.model.api import ApiSchema
 from tests.mock.mock import test_client, env, headers
